# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.oas36ty.test:8000/v1",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/users/{userID}/deactivate": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Deactivate User",
                "description": "Deactivate User",
                "operationId": "deactivateUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/tenant--header"
                    },
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error Message!"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "INVALID"
                                        },
                                        "errors": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected user_id is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
            "tenant--header": {
                "name": "X-Tenant",
                "in": "header",
                "description": "Tenant ID",
                "schema": {
                    "type": "string",
                    "default": "oas36ty"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Execute Login or Signup APIs to get the authentication token",
                "name": "Token Based Authentication",
                "scheme": "bearer"
            }
        }
    }
}